Address review: scope model default, gate composable researcher, fail-closed review policy, fix lint
Review fixes for PR #720:

- agency.json restored to the repo default (gpt-4o-mini); the experiment
  pins claude-sonnet-4-6 per-call in foo.agency instead, so tests and
  other runs keep the stable default.
- Composable researcher routes hosted search through hostedSearchTools()
  and drops the pinned provider, so it respects the model gate this PR
  adds and the caller's model selection. Adds the missing string return
  type. promptRewriter is out of the call path: it was an unconditional
  LLM round trip before any research; callers who want rewriting compose
  it explicitly.
- foo.agency review() inverts its warning policy: benign codes (AG3009)
  are dropped via the new typecheck ignoreCodes parameter, and any other
  warning rejects the program. Fails closed on future warning codes and
  gives ignoreCodes its production consumer.
- std-syntax-theme test updated to the ThemeName contract: object themes
  and unknown scheme names are now compile-time rejections, so the
  runtime probes for them are gone.
- stdlib/agency.agency formatter churn reverted; the diff there is now
  only the typecheck ignoreCodes change. Docs regenerated.
- Fix CI lint: warnOnOversizedToolSchemas moved from prompt.ts (which
  crossed the 1250-line structural cap) into toolSchemaSize.ts, typed
  against the minimal context slice it uses.
