+++ packages/agency-lang/docs/site/stdlib/agents/composable/researcher.md
@@ -18,12 +18,12 @@ Return the researcher's tools: the encyclopedia and fetch tools that need
 
 **Returns:** `any[]`
 
-([source](https://github.com/egonSchiele/agency-lang/tree/main/packages/agency-lang/stdlib/agents/composable/researcher.agency#L10))
+([source](https://github.com/egonSchiele/agency-lang/tree/main/packages/agency-lang/stdlib/agents/composable/researcher.agency#L8))
 
 ### researcherAgent
 
 ```ts
-researcherAgent(topic: string)
+researcherAgent(topic: string): string
 ```
 
 Research a topic in depth and give a response.
@@ -34,4 +34,6 @@ Research a topic in depth and give a response.
 |---|---|---|
 | topic | `string` |  |
 
-([source](https://github.com/egonSchiele/agency-lang/tree/main/packages/agency-lang/stdlib/agents/composable/researcher.agency#L40))
+**Returns:** `string`
+
+([source](https://github.com/egonSchiele/agency-lang/tree/main/packages/agency-lang/stdlib/agents/composable/researcher.agency#L38))
