+++ packages/agency-lang/stdlib/agents/composable/researcher.agency
@@ -1,9 +1,7 @@
-import { searchTools } from "std::agents/lib/search"
+import { hostedSearchTools, searchTools } from "std::agents/lib/search"
 import { webTools } from "std::agents/lib/toolkits"
 import { agentSkill } from "std::skills"
 import { systemMessage } from "std::thread"
-
-import { promptRewriter } from "./promptRewriter.agency"
 
 static const skills = agentSkill("researcher")
 
@@ -37,20 +35,17 @@ static const systemPrompt = """You are a research analyst. Your job is to answer
 - If your tools cannot retrieve the needed information, say so plainly and report what you tried. A partial answer with honest gaps beats a complete-looking answer with guesses.
 """
 
-export def researcherAgent(topic: string) {
+export def researcherAgent(topic: string): string {
   """
   Research a topic in depth and give a response.
   """
 
-  const rewrittenPrompt = promptRewriter("Please research this topic: ${topic}")
-
   thread {
     systemMessage(systemPrompt)
     const result = llm(
-      rewrittenPrompt,
+      "Please research this topic: <topic>${topic}</topic>",
       tools: [...buildTools()],
-      hostedTools: ["web_search"],
-      provider: "openai-responses",
+      hostedTools: hostedSearchTools(),
     )
     return result
   }
