+++ packages/agency-lang/tests/agency-js/stdlib/std-syntax-theme/fixture.json
@@ -1,6 +1,4 @@
 {
   "named": { "monokaiKeyword": true },
-  "custom": { "redKeyword": true },
-  "diff": { "monokaiKeyword": true, "hasRedBg": true, "hasGreenBg": true },
-  "bad": { "isFailure": true }
+  "diff": { "monokaiKeyword": true, "hasRedBg": true, "hasGreenBg": true }
 }
